[refactor] refactor ocr, ai-agent logic, improve maintainability and scalability
