+++ document-processing-service/src/main/java/com/microservice/document_processing_service/service/ai/AiClassifier.java
@@ -1,0 +1,9 @@
+package com.microservice.document_processing_service.service.ai;
+
+import com.microservice.document_processing_service.model.dto.TransactionItemDto;
+import java.util.List;
+import java.util.UUID;
+
+public interface AiClassifier {
+    List<TransactionItemDto> classifyItems(String ocrText, UUID documentId);
+}
