+++ document-processing-service/src/main/java/com/microservice/document_processing_service/service/core/DocumentStorageServiceImpl.java
@@ -1,4 +1,4 @@
-package com.microservice.document_processing_service.service.impl;
+package com.microservice.document_processing_service.service.core;
 
 import com.microservice.document_processing_service.service.DocumentStorageService;
 import io.minio.MinioClient;
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.UUID;
 
 @Service
