+++ document-processing-service/src/main/java/com/microservice/document_processing_service/service/ocr/FileDownloader.java
@@ -1,0 +1,52 @@
+package com.microservice.document_processing_service.service.ocr;
+
+import com.microservice.document_processing_service.exception.DocumentProcessingException;
+import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
+import io.minio.GetObjectArgs;
+import io.minio.MinioClient;
+import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.util.UUID;
+
+@Component
+@RequiredArgsConstructor
+public class FileDownloader {
+    private static final Logger logger = LoggerFactory.getLogger(FileDownloader.class);
+
+    private final MinioClient minioClient;
+
+    @Value("${minio.bucket}")
+    private String bucketName;
+
+    @CircuitBreaker(name = "minio-cb", fallbackMethod = "fallbackDownload")
+    public Path downloadFile(String objectName, UUID documentId) {
+        try {
+            Path tempFile = Files.createTempFile("ocr-", objectName.substring(objectName.lastIndexOf(".")));
+            try (InputStream inputStream = minioClient.getObject(
+                    GetObjectArgs.builder()
+                            .bucket(bucketName)
+                            .object(objectName)
+                            .build())) {
+                Files.copy(inputStream, tempFile, StandardCopyOption.REPLACE_EXISTING);
+            }
+            logger.info("File downloaded from MinIO: {} for document: {}", objectName, documentId);
+            return tempFile;
+        } catch (Exception e) {
+            logger.error("Failed to download file: {} for document: {}", objectName, documentId, e);
+            throw new DocumentProcessingException("Failed to download file from MinIO for document: " + documentId, e);
+        }
+    }
+
+    private Path fallbackDownload(String objectName, UUID documentId, Throwable t) {
+        logger.warn("Fallback triggered for file download for document {} due to: {}", documentId, t.getMessage());
+        throw new DocumentProcessingException("File download failed for document " + documentId + ": " + t.getMessage(), t);
+    }
+}
