+++ document-processing-service/src/main/java/com/microservice/document_processing_service/service/ocr/OcrProcessor.java
@@ -1,0 +1,7 @@
+package com.microservice.document_processing_service.service.ocr;
+
+import java.util.UUID;
+
+public interface OcrProcessor {
+    String extractTextFromImage(String objectName, UUID documentId);
+}
